+++ src/test/java/org/cojen/maker/AccessTest.java
@@ -17,6 +17,7 @@
 package org.cojen.maker;
 
 import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
 
 import java.lang.reflect.InvocationTargetException;
 
@@ -74,5 +75,15 @@ public class AccessTest {
         } catch (NoSuchMethodException e) {
             // Expected.
         }
+    }
+
+    @Test
+    public void lookup() throws Exception {
+        // Can call a private method when calling finishLookup.
+        ClassMaker cm = ClassMaker.begin("a.b.c.Dee").public_();
+        MethodMaker mm = cm.addMethod(null, "run").private_().static_();
+        MethodHandles.Lookup lookup = cm.finishLookup();
+        var clazz = lookup.lookupClass();
+        lookup.findStatic(clazz, "run", MethodType.methodType(void.class));
     }
 }